Add files via upload
Hello! I would like your help to correct this code I found online (and presented in the last meeting). By now, the two major errors I found is that this doesn't decrypt (i.e., it goes always in the same "direction") and it gives an error with some symbols (as, for example, with 1oz*).